Try caching .wp-cli/cache.

+++ utils/make-phar.php
@@ -70,7 +70,9 @@ function set_file_contents( $phar, $path, $content ) {
 	$phar[ $key ] = $content;
 }
 
-unlink( DEST_PATH );
+if ( file_exists( DEST_PATH ) ) {
+	unlink( DEST_PATH );
+}
 $phar = new Phar( DEST_PATH, 0, 'wp-cli.phar' );
 
 $phar->startBuffering();
